refactor: 헤더의 유저 정보를 가져오는클라이언트 세션 로직 서버세션으로 변경
